remove the user model and only use parseuser instead
