+++ app/src/main/java/com/example/womenwhocode/womenwhocode/models/PersonalizationDetail.java
@@ -2,6 +2,7 @@ package com.example.womenwhocode.womenwhocode.models;
 
 import com.parse.ParseClassName;
 import com.parse.ParseObject;
+import com.parse.ParseUser;
 
 import org.json.JSONArray;
 
@@ -17,12 +18,12 @@ public class PersonalizationDetail extends ParseObject {
         put(ANSWERS_KEY, answers);
     }
 
-    public void setUser(User user) {
+    public void setUser(ParseUser user) {
         put(USER_KEY, user);
     }
 
-    public User getUser()  {
-        return (User) getParseObject(USER_KEY);
+    public ParseUser getUser()  {
+        return (ParseUser) getParseObject(USER_KEY);
     }
 
     // parse array data type is json array
